// admin/src/types/banner.tsx
export interface BannerState {
    _id: string
    title: string
    banner: string
    isActive : boolean
}


export interface FloatingBannerState {
    _id: string
    title: string
    description : string
    imageUrl: string
    isActive : boolean
    createdAt : string
    updatedAt : string
}  